% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayNew.R
\name{DayNew}
\alias{DayNew}
\title{Matrix of the new cases/deathes/recovered/tested}
\usage{
DayNew(province,date)
}
\arguments{
\item{province}{province the province of Canada (List: "Nunavut","Northwest Territories", "Quebec","Newfoundland and Labrador",
"New Brunswick", "Nova Scotia", "Prince Edward Island", "Yukon", "Ontario", "Manitoba", "Saskatchewan", "    Alberta", "British Columbia") Double quotes are required.}

\item{date}{"yyyy-mm-dd" Double quotes are required. (eg. "2020-06-21")}
}
\value{
A matrix of new cases/deathes/recovered/tested of a chosen day in a chosen province of Canada.
}
\description{
Showing a matrix of the new cases/deathes/recovered/tested of a chosen day in a chosen province of Canada.
}
\examples{
DayNew("Ontario","2020-06-21")
}
\author{
Yujie & Jiahui
}
