% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshinynccp.R
\name{Inta_ncpp}
\alias{Inta_ncpp}
\title{Launch the function of Inta_ncpp}
\usage{
Inta_ncpp()
}
\value{
A shiny app would open
}
\description{
This function can launch a shiny app which is a interactive line plot of new case
in each province in Canada by day. You can choose a date in the slider and a province
in the menu below the slider. Then the plot will show you the new cases
by day in that province from the first recorded day to the day you choose
}
\author{
Yujie & Jiahui
}
