% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cum.R
\name{map_cum}
\alias{map_cum}
\title{Map of cumulative cases}
\usage{
map_cum(date)
}
\arguments{
\item{date}{"yyyy-mm-dd" Double quotes are required. (eg. "2020-06-21")}
}
\value{
Canada map with cumulative cases of a chosen day
}
\description{
Showing the cumulative cases of a chosen day on the Canada map.
}
\examples{
map_cum("2020-06-21")
}
\author{
Yujie & Jiahui
}
