% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_new.R
\name{map_new}
\alias{map_new}
\title{Map of new cases}
\usage{
map_new(date)
}
\arguments{
\item{date}{"yyyy-mm-dd" Double quotes are required. (eg. "2020-06-21")}
}
\value{
Canada map with new cases of a chosen day
}
\description{
Showing the new cases of a chosen day on the Canada map.
}
\examples{
map_new("2020-09-18")
}
\author{
Yujie & Jiahui
}
