% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ca.R
\name{plot_Ca}
\alias{plot_Ca}
\title{New cases/New deaths/Cumulative cases by day in Canada}
\usage{
plot_Ca(stopdate,type)
}
\arguments{
\item{stopdate}{"yyyy-mm-dd" Double quotes are required. (eg. "2020-06-21")}

\item{type}{"New", "Deaths", "Cumulative" Double quotes are required.}
}
\value{
A line chart of the new cases/new deaths/cumulative cases from the day the first case occured to a chosen day in Canada.
}
\description{
Show the new cases/new deaths/cumulative cases from the day the first case occured to a chosen day in Canada.
}
\examples{
plot_Ca("2020-06-12","New")
plot_Ca("2020-09-12","Cumulative")
plot_Ca("2020-11-12","Deaths")
}
\author{
Yujie & Jiahui
}
