% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Prov.R
\name{plot_Prov}
\alias{plot_Prov}
\title{New cases/New deaths/Cumulative cases by day in Canada province}
\usage{
plot_Prov(province,stopdate,type)
}
\arguments{
\item{province}{the province of Canada (List: "Nunavut","Northwest Territories", "Quebec","Newfoundland and Labrador",
"New Brunswick", "Nova Scotia", "Prince Edward Island", "Yukon", "Ontario", "Manitoba", "Saskatchewan", "    Alberta", "British Columbia") Double quotes are required.}

\item{stopdate}{"yyyy-mm-dd" Double quotes are required. (eg. "2020-06-21")}

\item{type}{"New", "Deaths", "Cumulative" Double quotes are required.}
}
\value{
A line chart of new cases/new deaths/cumulative cases from the day the first case occured to a chosen day in the chosen Canada province.
}
\description{
Show the new cases/new deaths/cumulative cases from the day the first case occured to a chosen day in the chosen Canada province.
}
\examples{
plot_Prov("Manitoba","2020-06-10","New")
plot_Prov("Alberta","2020-09-10","Deaths")
plot_Prov("British Columbia","2020-12-10","Cumulative")
}
\author{
Yujie & Jiahui
}
